% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_deck.R
\name{tm_deck}
\alias{tm_deck}
\title{Deck mode options}
\usage{
tm_deck(pitch, basemap.show)
}
\arguments{
\item{pitch}{The pitch angle. By default 60}

\item{basemap.show}{Show basemap? By default TRUE}
}
\description{
Deck mode options. These options are specific to the deck mode.
}
\examples{
library(tmap)

tmap_mode("deck")
tm_shape(World) + tm_polygons("HPI", fill.scale = tm_scale_intervals(values = "brewer.rd_yl_gn")) + tm_deck(pitch = 60, basemap.show = FALSE)
}
