% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{tmap.deckgl-package}
\alias{tmap.deckgl}
\alias{tmap.deckgl-package}
\title{Extension for tmap: maps are shown via deck.gl}
\description{
A new tmap mode, enabled via tmap_mode("deck").
}
\examples{
library(tmap)

tmap_mode("deck")
tm_shape(World) + tm_polygons("HPI", fill.scale = tm_scale_intervals(values = "brewer.rd_yl_gn")) + tm_deck(pitch = 60, basemap.show = FALSE)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-tmap/tmap.deckgl}
  \item Report bugs at \url{https://github.com/r-tmap/deckgl/issues}
}

}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\concept{GIS}
\concept{bubble map}
\concept{choropleth}
\concept{statistical maps}
\concept{thematic maps}
