% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmapDeckGL.R, R/tmapDeckGLAux.R,
%   R/tmapDeckGLComp_charts.R, R/tmapDeckGLInit.R, R/tmapDeckGLLegend.R,
%   R/tmapDeckGLRun.R, R/tmapDeckGLShape.R, R/tmapDeckGLXYlab.R,
%   R/tmapDeckGL_layers.R
\name{tmapDeckGLWrap}
\alias{tmapDeckGLWrap}
\alias{tmapDeckGLXtab}
\alias{tmapDeckGLTilesPrep}
\alias{tmapDeckGLTiles}
\alias{tmapDeckGLGridPrep}
\alias{tmapDeckGLGrid}
\alias{tmapDeckGLGridXLab}
\alias{tmapDeckGLGridYLab}
\alias{tmapDeckGLCompPrepare}
\alias{tmapDeckGLCompHeight}
\alias{tmapDeckGLCompWidth}
\alias{tmapDeckGLLegPlot}
\alias{tmapDeckGLInit}
\alias{tmapDeckGLAux}
\alias{tmapDeckGLLegend}
\alias{tmapDeckGLRun}
\alias{tmapDeckGLShape}
\alias{tmapDeckGLOverlay}
\alias{tmapDeckGLXlab}
\alias{tmapDeckGLYlab}
\alias{tmapDeckGLPolygons}
\alias{tmapDeckGLLines}
\alias{tmapDeckGLSymbols}
\alias{tmapDeckGLRaster}
\alias{tmapDeckGLText}
\title{Internal tmap methods}
\usage{
tmapDeckGLWrap(label, facet_row, facet_col, facet_page, o)

tmapDeckGLXtab(label, facet_row, facet_col, facet_page, o)

tmapDeckGLTilesPrep(a, bs, id, o)

tmapDeckGLTiles(bi, bbx, facet_row, facet_col, facet_page, id, pane, group, o)

tmapDeckGLGridPrep(a, bs, id, o)

tmapDeckGLGrid(bi, bbx, facet_row, facet_col, facet_page, id, pane, group, o)

tmapDeckGLGridXLab(bi, bbx, facet_row, facet_col, facet_page, o)

tmapDeckGLGridYLab(bi, bbx, facet_row, facet_col, facet_page, o)

tmapDeckGLCompPrepare(comp, o)

tmapDeckGLCompHeight(comp, o)

tmapDeckGLCompWidth(comp, o)

tmapDeckGLLegPlot(comp, deck, o)

tmapDeckGLInit(o, return.asp = FALSE, vp)

tmapDeckGLAux(o, q)

tmapDeckGLLegend(
  comp,
  o,
  facet_row = NULL,
  facet_col = NULL,
  facet_page,
  class,
  stack,
  stack_auto,
  pos.h,
  pos.v,
  bbox
)

tmapDeckGLRun(o, show, knit, args)

tmapDeckGLShape(bbx, facet_row, facet_col, facet_page, o)

tmapDeckGLOverlay(facet_row, facet_col, facet_page, o)

tmapDeckGLXlab(facet_page, o)

tmapDeckGLYlab(facet_page, o)

tmapDeckGLPolygons(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapDeckGLLines(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapDeckGLLines(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o
)

tmapDeckGLSymbols(
  shpTM,
  dt,
  pdt,
  popup.format,
  hdt,
  idt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapDeckGLRaster(
  shpTM,
  dt,
  gp,
  pdt,
  popup.format,
  hdt,
  idt,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)

tmapDeckGLText(
  shpTM,
  dt,
  gp,
  bbx,
  facet_row,
  facet_col,
  facet_page,
  id,
  pane,
  group,
  o,
  ...
)
}
\arguments{
\item{label}{label}

\item{facet_row, facet_col, facet_pace}{row column and page id}

\item{o}{the list of options}

\item{bs}{bs}

\item{id}{id}

\item{bi}{bi}

\item{bbx}{bbx}

\item{pane}{pane}

\item{group}{group}

\item{comp}{the shape object}

\item{deck}{deck object}

\item{return.asp}{return.asp}

\item{vp}{vp}

\item{q}{q}

\item{class}{class}

\item{stack}{stack}

\item{stack_auto}{stack_auto}

\item{pos.h}{pos.h}

\item{pos.v}{pos.v}

\item{bbox}{bbox}

\item{show}{show}

\item{knit}{knit}

\item{shpTM, }{dt, pdt, popup.format, hdt, idt, gp args}

\item{arg}{args}
}
\description{
Internal tmap methods
}
\keyword{internal}
